# Add cycle length cache for 7.8x speedup loading branches