--- a/parser-typechecker/src/Unison/Codebase/Branch.hs
+++ b/parser-typechecker/src/Unison/Codebase/Branch.hs
@@ -510,13 +510,13 @@ numHashChars _b = 3
 
 -- This type is a little ugly, so we wrap it up with a nice type alias for
 -- use outside this module.
-type Cache m = Cache.Cache m (Causal.RawHash Raw) (UnwrappedBranch m)
+type Cache m = Cache.Cache (Causal.RawHash Raw) (UnwrappedBranch m)
 
-boundedCache :: MonadIO m => Word -> m (Cache m)
+boundedCache :: MonadIO m => Word -> m (Cache m2)
 boundedCache = Cache.semispaceCache
 
 -- Can use `Cache.nullCache` to disable caching if needed
-cachedRead :: forall m . Monad m
+cachedRead :: forall m . MonadIO m
            => Cache m
            -> Causal.Deserialize m Raw Raw
            -> (EditHash -> m Patch)