@aryairani 2021-05-18
The average number of lookups per defn was 41.7